class activity to create methods to return square of numbers and 2D array

+++ java/JAVA/src/ActivityOne.java
@@ -1,5 +1,5 @@
 public class ActivityOne {
-    static int sqr(int a){
+    static double sqr(double a){
         return a * a;
     }
